Core/Misc: Fix some -Wconversion warnings
Conflicts:
	src/server/collision/DynamicTree.cpp
	src/server/collision/Models/GameObjectModel.cpp
	src/server/collision/RegularGrid.h
	src/server/game/Battlefield/Battlefield.cpp
	src/server/game/Battlegrounds/Zones/BattlegroundSA.cpp
	src/server/game/Entities/Object/Object.cpp
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Unit/StatSystem.cpp
	src/server/game/Entities/Unit/Unit.cpp
	src/server/scripts/EasternKingdoms/Karazhan/boss_netherspite.cpp
	src/server/scripts/EasternKingdoms/Karazhan/boss_nightbane.cpp
	src/server/scripts/EasternKingdoms/Karazhan/bosses_opera.cpp
	src/server/scripts/EasternKingdoms/SunwellPlateau/boss_kiljaeden.cpp
	src/server/scripts/EasternKingdoms/ZulAman/boss_janalai.cpp
	src/server/scripts/Kalimdor/TempleOfAhnQiraj/boss_cthun.cpp
	src/server/scripts/Kalimdor/TempleOfAhnQiraj/boss_viscidus.cpp
	src/server/scripts/Northrend/AzjolNerub/Ahnkahet/boss_prince_taldaram.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheChampion/trial_of_the_champion.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheCrusader/boss_twin_valkyr.cpp
	src/server/scripts/Northrend/FrozenHalls/ForgeOfSouls/boss_devourer_of_souls.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_icecrown_gunship_battle.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_lord_marrowgar.cpp
	src/server/scripts/Northrend/Naxxramas/boss_heigan.cpp
	src/server/scripts/Northrend/Nexus/EyeOfEternity/boss_malygos.cpp
	src/server/scripts/Northrend/Nexus/Oculus/boss_varos.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_general_vezax.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_hodir.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_kologarn.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_yogg_saron.cpp
	src/server/scripts/Northrend/zone_grizzly_hills.cpp
	src/server/scripts/Northrend/zone_sholazar_basin.cpp
	src/server/scripts/Northrend/zone_zuldrak.cpp
	src/server/scripts/Outland/BlackTemple/boss_shade_of_akama.cpp
	src/server/scripts/Outland/CoilfangReservoir/SerpentShrine/boss_leotheras_the_blind.cpp
	src/server/scripts/Outland/TempestKeep/Eye/boss_astromancer.cpp
	src/server/scripts/Outland/TempestKeep/botanica/boss_warp_splinter.cpp
	src/server/shared/Utilities/Timer.h

+++ src/server/shared/Utilities/Util.h
@@ -623,7 +623,7 @@ class EventMap
             if (!phase)
                 _phase = 0;
             else if (phase <= 8)
-                _phase = (1 << (phase - 1));
+                _phase = uint8(1 << (phase - 1));
         }
 
         /**
@@ -634,7 +634,7 @@ class EventMap
         void AddPhase(uint8 phase)
         {
             if (phase && phase <= 8)
-                _phase |= (1 << (phase - 1));
+                _phase |= uint8(1 << (phase - 1));
         }
 
         /**
@@ -645,7 +645,7 @@ class EventMap
         void RemovePhase(uint8 phase)
         {
             if (phase && phase <= 8)
-                _phase &= ~(1 << (phase - 1));
+                _phase &= uint8(~(1 << (phase - 1)));
         }
 
         /**
